+++ data/chill-musics-meu.csv
@@ -1,0 +1,34 @@
+danceability,energy,key,loudness,mode,speechiness,acousticness,instrumentalness,liveness,valence,tempo,duration_ms,genres__001,genres__002,genres__003,genres__004,genres__005,genres__006,genres__007,genres__008,genres__009,genres__010
+"509,000","678,000",1,"-7536,000",0,"157,000",0.64,"737,000","132,000","319,000","129665,000",199615,catstep,edm,electro house,,,,,,,
+"956,000","373,000",6,"-7963,000",1,0.0773,"569,000","394,000","314,000","568,000","119994,000",268333,chillwave,indie pop,indie rock,indie soul,indietronica,kiwi rock,neo-psychedelic,new rave,portland hip hop,psychedelic pop
+"341,000","354,000",2,"-15346,000",1,0.0846,"478,000",0.0782,"155,000","117,000","124872,000",206160,electropop,indie anthem-folk,indie cafe pop,indie folk,vapor soul,,,,,
+"684,000","255,000",10,"-9381,000",0,0.0445,"805,000",0.79,0.0884,0.33,"84226,000",152000,lo-fi beats,chillhop,jazz boom bap,,,,,,,
+"783,000","467,000",6,"-9474,000",1,0.0503,"729,000","809,000","109,000","236,000","90838,000",237793,new french touch,tropical house,,,,,,,,
+"786,000","686,000",4,-7.18,1,0.0289,"276,000","668,000",0.0416,"542,000","119997,000",261773,la indie,,,,,,,,,
+"772,000","745,000",11,"-10194,000",1,0.0405,"595,000",0.87,"117,000","166,000","116014,000",273162,new french touch,tropical house,,,,,,,,
+"765,000",0.18,10,"-10989,000",1,0.0788,"656,000","813,000","128,000","493,000","80043,000",150000,chillhop,lo-fi beats,,,,,,,,
+"499,000","269,000",9,"-12692,000",1,0.0625,"693,000","799,000",0.0638,0.0422,"152134,000",123288,,,,,,,,,,
+"791,000","267,000",5,"-11997,000",1,0.0415,"674,000","534,000","102,000","544,000",165.99,215111,indie pop,,,,,,,,,
+"533,000","545,000",7,"-7681,000",1,0.0243,"207,000","797,000","113,000","419,000","98014,000",270918,alternative rock,art pop,dream pop,indie rock,nu gaze,reading indie,shoegaze,slow core,,
+"632,000","396,000",2,"-12926,000",0,0.0324,0.11,"827,000","101,000","885,000","81734,000",221219,funk,neo-psychedelic,soul,,,,,,,
+"542,000","518,000",1,"-11332,000",0,0.0313,"224,000","858,000",0.0797,"531,000","85922,000",163962,funk,neo-psychedelic,soul,,,,,,,
+"601,000","499,000",6,"-11131,000",1,0.0343,"815,000","712,000","109,000","531,000","96981,000",242860,electropop,indie poptimism,indietronica,new french touch,pop,tropical house,vapor soul,vapor pop,,
+"788,000","711,000",11,"-9564,000",1,0.0694,"244,000",2.1E-06,"223,000",0.13,"113048,000",224105,new french touch,,,,,,,,,
+0.87,"572,000",6,"-7667,000",1,"39,000",0.16,0.00282,0.0993,"854,000",119.98,198582,australian pop,pop quebecois,,,,,,,,
+"498,000","545,000",7,"-7917,000",0,0.0481,0.0216,"808,000",0.0996,0.0383,"177984,000",248785,bass trap,brostep,edm,electro house,electronic trap,pop edm,,,,
+"804,000","577,000",10,"-14929,000",1,0.14,"922,000","939,000","113,000","181,000","79486,000",150967,japanese chillhop,lo-fi beats,,,,,,,,
+"702,000","629,000",6,"-6404,000",0,0.0274,"474,000","908,000","69,000","104,000","100918,000",217195,wave,,,,,,,,,
+"454,000","401,000",3,"-10255,000",1,"26,000","131,000","845,000","137,000","321,000","93719,000",320030,ambient pop,dream pop,el paso indie,shoegaze,,,,,,
+"737,000","284,000",8,"-8535,000",0,"114,000","381,000",0.89,"812,000",0.75,"87967,000",143437,chillhop,lo-fi beats,,,,,,,,
+"839,000","363,000",6,"-5819,000",0,0.0591,"869,000",0.85,"104,000",0.84,"90052,000",117420,ambeat,chillhop,jazz boom bap,,,,,,,
+"892,000",0.38,8,"-5847,000",1,"272,000","308,000","586,000","117,000",0.62,"94029,000",226693,ambeat,chillhop,jazz boom bap,,,,,,,
+"382,000",0.54,4,"-8768,000",1,0.0289,"908,000","651,000",0.0915,0.0948,"139693,000",212963,indie electropop,indie poptimism,vapor soul,,,,,,,
+"691,000",0.49,11,"-12883,000",0,"275,000","918,000","376,000",0.0751,0.0385,"107965,000",244506,new french touch,vapor soul,,,,,,,,
+"549,000","594,000",8,"-6302,000",1,0.0496,0.65,3.57E-05,"127,000","324,000","99936,000",236279,deep house,deep tropical house,indie poptimism,new french touch,tropical house,vapor soul,,,,
+"715,000","522,000",4,"-14952,000",0,0.0546,"366,000","834,000",0.0877,0.0384,"107977,000",280059,new french touch,vapor soul,,,,,,,,
+"737,000","528,000",3,"-9583,000",0,0.0981,"767,000",4.92E-06,0.0782,"332,000","126006,000",206635,vapor soul,,,,,,,,,
+"729,000",0.72,3,-6.69,1,0.0371,"125,000","353,000","178,000",0.0388,"122015,000",578041,australian electropop,indietronica,,,,,,,,
+"784,000","312,000",4,"-14648,000",0,0.0513,"576,000","851,000","128,000",0.63,129.94,147692,chillhop,lo-fi beats,,,,,,,,
+"616,000","375,000",5,"-16921,000",1,0.0405,"905,000","896,000","332,000","333,000","159972,000",132000,chillhop,lo-fi beats,,,,,,,,
+"718,000","402,000",2,"-14109,000",1,0.0398,0.0177,0.68,0.0907,"95,000","121995,000",353227,deep disco house,deep euro house,electronica,french tech house,house,microhouse,minimal tech house,minimal techno,organic house,tech house
+"717,000","193,000",9,"-16141,000",0,0.0459,"802,000","893,000",0.0653,0.0447,"109992,000",195136,,,,,,,,,,
