+++ data/resultados_teste.txt
@@ -1,0 +1,44 @@
+BASE NORMAL:
+
+danceability: 0.144
+energy: 0.155
+loudness: 3.170
+speechiness: 0.060
+acousticness: 0.284
+instrumentalness: 0.301
+liveness: 0.174
+valence: 0.266
+tempo: 25.008
+duration_ms: 83206.615
+
+APÓS ADIÇÃO DE MÚSICA ROCK INDIE:
+
+danceability: 0.145
+energy: 0.176
+loudness: 3.420
+speechiness: 0.059
+acousticness: 0.290
+instrumentalness: 0.315
+liveness: 0.172
+valence: 0.275
+tempo: 25.074
+duration_ms: 82159.043
+
+APÓS ADIÇÃO DE MÚSICA CHILL + REMOÇÃO DA MÚSICA ROCK INDIE (OBVIAMENTE):
+
+danceability: 0.144
+energy: 0.159
+loudness: 3.178
+speechiness: 0.060
+acousticness: 0.286
+instrumentalness: 0.308
+liveness: 0.171
+valence: 0.268
+tempo: 24.913
+duration_ms: 82396.976
+
+Perguntas:
+
+Para adicionar uma musica na playlist, essa musica tem que estar dentro do range de cada atributo?
+Ou só dos atributos com peso maior/ranking maior?
+Ou a gente vai fazer uma média dos ranges/atributos/dp?...
