+++ env.py
@@ -1,0 +1,4 @@
+import os
+
+# Set environment variables
+os.environ['CLIENT_ID'] = '648f3dc76d64441abf6a5fc53fcdd843'
