+++ lib/metadata/metadataUtils.py
@@ -1,0 +1,6 @@
+import statistics
+
+def recalculateDP(array: [], value: float) -> float:
+    arr = array.copy()
+    arr.append(value)
+    return float("{:.3f}".format(statistics.pstdev(arr)))
