--- lib/spotify/utils.py
@@ -1,2 +1,0 @@
-def getTokenFromURL(url):
-    return str(url).split('#')[1].split('&')[0].split('=')[1]
